refactor(architecture): dissolve three import-cycle clusters via leaf extractions
Pays down three of the five ALLOWED_CYCLES clusters baselined in #1358,
using the extract-shared-symbol-to-a-leaf patterns demonstrated there:

- requirements-manager/checks (7 files): the only back-edges were
  `import type { CheckResultError }` from the check router. The type now
  lives in the Tier-0 `types/requirements.types.ts` leaf; the router
  re-exports it so the barrel surface is unchanged.
- components/interactive-tutorial (4 files): `use-section-requirements`
  needed only the `DEFAULT_INTERACTIVE_SECTION_TITLE` constant (moved to
  `section-titles.ts`), and `section-numbering` imported
  `InteractiveConditional` solely for an identity check — replaced with
  a self-tagging marker (`skip-section-numbering.ts`), which also removes
  that component from the cycle-load-order hazard noted in the file header.
- components/docs-panel (4 files): both children imported the
  `CombinedLearningJourneyPanel` class as a prop type. They now type
  against `DocsPanelModelOperations` (extended with the two alignment
  methods the class already exposes publicly).

Part of #1359 — the telemetry/OpenFeature and interactive-engine <->
requirements-manager clusters remain.
