+++ src/components/interactive-tutorial/interactive-section.tsx
@@ -105,6 +105,7 @@ import {
 } from '../../global-state/completion-store';
 import { dispatchProgress } from '../../global-state/progress-events';
 import { computeCursor, deriveSectionState, initialSectionState, sectionReducer } from './section-state';
+import { DEFAULT_INTERACTIVE_SECTION_TITLE, PASSIVE_SECTION_TITLE } from './section-titles';
 import {
   getDocumentStepPosition,
   nextSectionCounter,
@@ -119,11 +120,9 @@ export {
   getDocumentStepPosition,
 } from '../../global-state/section-registry';
 
-// Interactive Section title fallback
-export const DEFAULT_INTERACTIVE_SECTION_TITLE = 'Interactive section';
-
-// Interactive Section title fallback for sections with no interactive steps (passive content only)
-export const PASSIVE_SECTION_TITLE = 'Steps';
+// Re-exports preserved for back-compat with tests and the barrel. New code
+// should import directly from `./section-titles`.
+export { DEFAULT_INTERACTIVE_SECTION_TITLE, PASSIVE_SECTION_TITLE } from './section-titles';
 
 // Reset every counter (registry + offsets + per-step-type anonymous-ID
 // counters). Called when new content loads. The registry's own state
