+++ src/components/interactive-tutorial/section-titles.ts
@@ -1,0 +1,5 @@
+// Interactive Section title fallback
+export const DEFAULT_INTERACTIVE_SECTION_TITLE = 'Interactive section';
+
+// Interactive Section title fallback for sections with no interactive steps (passive content only)
+export const PASSIVE_SECTION_TITLE = 'Steps';
